backend/requirements.txt
# ── API / HTTP ────────────────────────────────────────────────────────────────
fastapi>=0.111.0
uvicorn[standard]>=0.29.0      # ASGI server (dev + Uvicorn worker for gunicorn)
gunicorn>=22.0.0               # Production process manager (Azure App Service)
httpx>=0.27.0                  # Async HTTP client — external reward program APIs,
                               # async FastAPI test client
python-multipart>=0.0.9        # Required by FastAPI for file upload / form data endpoints

# ── Database ──────────────────────────────────────────────────────────────────
sqlalchemy>=2.0.30             # ORM + async session management
asyncpg>=0.29.0                # Async PostgreSQL driver (used by SQLAlchemy)
greenlet>=3.0.0                # Required by SQLAlchemy async internals
alembic>=1.13.1                # Schema migrations
psycopg2-binary>=2.9.9         # Sync PostgreSQL driver — required by Alembic CLI
                               # and useful for one-off scripts / data loads

# ── Data validation & settings ────────────────────────────────────────────────
pydantic>=2.7.0                # Request/response schema validation
pydantic-settings>=2.2.1       # Typed settings/config from env vars (BaseSettings)

# ── Optimization ──────────────────────────────────────────────────────────────
scipy>=1.13.0                  # linprog: optimal cross-card allocation when capped
                               # multiplier groups bind (overflow reallocation in
                               # the calculator's segmented earn path)

# ── Serialization ─────────────────────────────────────────────────────────────
orjson>=3.10.0                 # Fast JSON serializer; FastAPI uses it automatically
                               # when installed (significant speedup for large responses)

# ── Authentication ────────────────────────────────────────────────────────────
google-auth>=2.29.0            # Verify Google OAuth2 ID tokens server-side
PyJWT>=2.8.0                   # Issue/verify stateless JWT session tokens

# ── Configuration ─────────────────────────────────────────────────────────────
python-dotenv>=1.0.1           # Load .env in local dev

# ── Azure ─────────────────────────────────────────────────────────────────────
azure-identity>=1.17.0         # Managed Identity + service principal auth;
                               # enables passwordless connection to Azure PostgreSQL
